requirements.txt
streamlit==0.92.0
opencv-python==4.5.3.56
face-recognition==1.3.0
matplotlib==3.4.3
numpy==1.21.1
Pillow==8.3.1